Add useWalletConnectManager and change transaction confirmation to drawer
