+++ packages/nextjs/components/Header.tsx
@@ -1,6 +1,8 @@
 "use client";
 
 import { RandomLoadingBackground } from "./RandomLoadingBackground";
+import { WalletConnectProposalDrawer } from "./burnerwallet/WalletConnectProposalDrawer";
+import { WalletConnectTransactionDrawer } from "./burnerwallet/WalletConnectTransactionDrawer";
 import Jazzicon, { jsNumberForAddress } from "react-jazzicon";
 import { useLocalStorage } from "usehooks-ts";
 import { useAccount, useNetwork, useSwitchNetwork } from "wagmi";
@@ -84,6 +86,8 @@ export const Header = ({ isGenerateWalletLoading, showIntro, updateHistory }: He
           <ReceiveDrawer address={connectedAddress} />
           <SendDrawer address={connectedAddress} updateHistory={updateHistory} />
           <WalletConnectDrawer />
+          <WalletConnectProposalDrawer />
+          <WalletConnectTransactionDrawer />
         </div>
       </div>
     </div>
